+++ ChangeLog.md
@@ -1,3 +1,9 @@
+# 0.1.2
+
+* Fix error handling. Kura::ApiError#reason was missing.
+* add Kura::Client#list_tabledata API.
+
 # 0.1.1
+
 * Add Kura::Client#tables API.
 * Kura::Client#wait_job yield block every second if block was passed.